adicionada rota, svg e imagem ao template de erro
