+++ .env
@@ -17,7 +17,7 @@ PYTHON_DEBUG=1
 FLASK_APPSETTINGS=settings.Dev
 FLASK_ENV=development
 FLASK_SECRET_KEY=[REPLACE KEY]
-FLASK_PORT=80
+FLASK_PORT=81
 
 # --- SQL Alchemy settings
 BASE_DB_URI=postgresql://postgres:{PASSWORD}@db:5432/store
