+++ docker-compose.yml
@@ -7,7 +7,7 @@ services:
       target: dellstore
     restart: always
     ports:
-      - 5432:5432
+      - 5431:5432
     environment:
       POSTGRES_DB: ${DB_NAME}
       POSTGRES_USER: ${DB_USER}
