+++ templates/404.jinja
@@ -1,0 +1,7 @@
+{% extends "_base.jinja" %}
+
+{% block CONTENT %}
+<a href="{{ url_for('signin') }}"> 
+    <img class="uk-align-center uk-margin-xlarge-top" alt="" height="50%" width="50%" src="{{ url_for('static', filename='img/error_404.svg') }}">
+</a>
+{% endblock %}
