+++ views/__init__.py
@@ -1,2 +1,3 @@
 from .account import * 
 from .eventhooks import *
+from .http_errors import *
