+++ views/http_errors.py
@@ -1,0 +1,8 @@
+from flask import render_template
+
+from app import app
+
+
+@app.errorhandler(404)
+def erro_404(error):
+    return render_template('404.jinja')
